+++ astroapp/application/pkg/sender/sendEvent.go
@@ -20,10 +20,6 @@ type EventSender interface {
 	SendEvent(event api.Event, appName string, side string, q queue.QueueInterface)
 }
 
-// BinaryEventSender handles binary events for PPXF
-type BinaryEventSender interface {
-	SendBinaryEvent(event api.BinaryEvent, appName string, side string, q queue.QueueInterface)
-}
 
 type RabbitMQSender struct {
 	Queue       queue.QueueInterface
@@ -39,11 +35,7 @@ func NewRabbitMQSender(queue queue.QueueInterface, utils common.UtilsInterface, 
 	}
 }
 
-// BinaryRabbitMQSender handles binary events via RabbitMQ
-type BinaryRabbitMQSender struct {
-	Queue queue.QueueInterface
-	Utils common.UtilsInterface
-}
+
 
 func (s *RabbitMQSender) SendEvent(event api.Event, appName string, side string, q queue.QueueInterface) {
 	var queueName string
@@ -109,52 +101,3 @@ func (s *RabbitMQSender) SendEvent(event api.Event, appName string, side string,
 	log.Printf("     Files: %s\n", strings.Join(filenames, ", "))
 }
 
-// SendBinaryEvent handles binary events for PPXF (preserves .fits file integrity)
-func (bs *BinaryRabbitMQSender) SendBinaryEvent(event api.BinaryEvent, appName string, side string, q queue.QueueInterface) {
-	var queueName string
-
-	if side == "producer" {
-		queueName = "producer_to_processor_queue"
-	} else {
-		queueName = "processor_to_producer_queue"
-	}
-
-	err := q.Connect()
-	if err != nil {
-		bs.Utils.FailOnError("Failed to connect to RabbitMQ", err)
-	}
-	defer q.Close()
-
-	err = q.DeclareQueue(queueName)
-	if err != nil {
-		bs.Utils.FailOnError(fmt.Sprintf("Failed to declare queue"), err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	eventJSON, err := json.Marshal(event)
-	if err != nil {
-		bs.Utils.FailOnError("Failed to marshal binary event", err)
-	}
-
-	headers := make(amqp.Table)
-	headers["batch_size"] = len(event.Files)
-	headers["app_name"] = appName
-	headers["event_id"] = event.ID
-	headers["is_binary"] = true // Flag to identify binary events
-
-	filenames := []string{}
-	for _, f := range event.Files {
-		filenames = append(filenames, f.Name)
-	}
-	headers["filenames"] = strings.Join(filenames, ",")
-
-	err = q.Publish(ctx, queueName, eventJSON, headers)
-	if err != nil {
-		bs.Utils.FailOnError("Failed to publish binary message: %v", err)
-	}
-	log.Printf("DEBUG: Published binary message to queue %s", queueName)
-	log.Printf(" [x] Sent binary batch with %d files for app %s\n", len(event.Files), appName)
-	log.Printf("     Binary Files: %s\n", strings.Join(filenames, ", "))
-}
